refactor: improve type safety and thread-safety in rpc_client and supervisor

+++ virus_executor_service/include/transport/ves_control_interface.h
@@ -15,7 +15,7 @@ constexpr int32_t VIRUS_PROTECTION_EXECUTOR_SA_ID = 1251;
 constexpr uint32_t VES_OPEN_SESSION_REQUEST_VERSION = 1;
 constexpr size_t VES_OPEN_SESSION_MAX_ENGINE_KINDS = 32;
 
-enum class VesEngineKind : uint32_t {
+enum class VesEngineKind : uint8_t {
     Scan = 1,
 };
