+++ virus_executor_service/include/ves/ves_protocol.h
@@ -11,7 +11,7 @@ enum class VesOpcode : uint8_t {
 
 inline constexpr uint32_t VES_EVENT_DOMAIN_RUNTIME = 1;
 
-enum class VesEventType : uint32_t {
+enum class VesEventType : uint8_t {
     RandomScanResult = 1,
     RandomHealthHint = 2,
     RandomLifecycle = 3,
